chore: renaming frontend landing page
