+++ app/controllers/api/v1/listing_interest_controller.rb
@@ -1,5 +1,6 @@
-class Api::V1::EmailController < ApiController
-  def confirmation
+# Backend API for are you interested email
+class Api::V1::ListingInterestController < ApiController
+  def index
     token = params[:t]
     secret = get_secret
 
@@ -24,17 +25,17 @@ class Api::V1::EmailController < ApiController
       end
 
       listing_id = listing_id_map(b)
-      redirect_to "/confirming_email?listing=#{listing_id}&response=#{r}"
+      redirect_to "/listing_interest?listing=#{listing_id}&response=#{r}"
     rescue JWT::ExpiredSignature
       Rails.logger.error('Token expired, not able to create fieldUpdateComment')
       decoded_expired_token = JWT.decode(token, secret, true,
                                          { verify_expiration: false })
       b = decoded_expired_token[0]['b']
       listing_id = listing_id_map(b)
-      redirect_to "/confirming_email?listing=#{listing_id}&response=x"
+      redirect_to "/listing_interest?listing=#{listing_id}&response=x"
     rescue StandardError => e
       Rails.logger.error("Error when creating fieldUpdateComment #{e}")
-      redirect_to "/confirming_email?listing=#{listing_id}&response=e"
+      redirect_to "/listing_interest?listing=#{listing_id}&response=e"
     end
   end
 
@@ -45,9 +46,9 @@ class Api::V1::EmailController < ApiController
   end
 
   def listing_id_map(listing_number)
-    array = ['a0W0P00000DZYzVUAX', 'a0W4U00000KnLRMUA3', 
-             'a0W4U00000IYEb4UAH', 'a0W4U00000IYSM4UAP',
-             'a0W4U00000Ih1V2UAJ', 'a0W4U00000KnCZRUA3']
+    array = %w[a0W0P00000DZYzVUAX a0W4U00000KnLRMUA3
+               a0W4U00000IYEb4UAH a0W4U00000IYSM4UAP
+               a0W4U00000Ih1V2UAJ a0W4U00000KnCZRUA3]
     array[listing_number]
   end
