--- app/views/confirming_email/confirming_email.html.slim
@@ -1,1 +1,0 @@
-== react_component 'ConfirmingEmail', props: @confirming_email_props
