+++ app/views/listing_interest_page/listing_interest.html.slim
@@ -1,0 +1,1 @@
+== react_component 'ListingInterestPage', props: @listing_interest_props
