+++ config/routes.rb
@@ -37,7 +37,7 @@ Rails.application.routes.draw do
           get 'eligibility' => 'listings#eligibility'
         end
       end
-      get 'trk' => 'email#confirmation'
+      get 'trk' => 'listing_interest#index'
       scope '/short-form' do
         post 'validate-household' => 'short_form#validate_household'
         get 'listing-application/:listing_id' => 'short_form#show_listing_application_for_user'
@@ -93,7 +93,7 @@ Rails.application.routes.draw do
   get '(:lang)/privacy' => 'assistance#privacy', lang: /(en|es|zh|tl)/
   get '(:lang)/disclaimer' => 'assistance#disclaimer', lang: /(en|es|zh|tl)/
 
-  get '(:lang)/confirming_email' => 'confirming_email#confirming_email', lang: /(en|es|zh|tl)/
+  get '(:lang)/listing_interest' => 'listing_interest_page#index', lang: /(en|es|zh|tl)/
 
   # fallback to Angular-only controller for all un-migrated pages.
   get '*path', to: 'angular#index', constraints: ->(req) { req.format == :html || req.format == '*/*' }
